move span factory into utils folder
